JUL7_LPC55S69EVK

+++ stream_nodes/AI/arm/detector2D/arm_stream_detector2D.c
@@ -108,7 +108,7 @@ void arm_stream_detector2D (int32_t command, stream_handle_t instance, stream_xd
                 memory pointers are in the same order as described in the SWC manifest
         */
         case STREAM_RESET: 
-        {   stream_al_services *stream_entry = (stream_al_services *)(uint64_t)data;
+        {   stream_al_services *stream_entry = (stream_al_services *)data;
             intPtr_t *memresults = (intPtr_t *)instance;
             uint16_t preset = RD(command, PRESET_CMD);
             uint8_t *pt8bdst, i, n;
