+++ stream_nodes/Audio/arm/decompressor/arm_stream_decompressor.c
@@ -81,7 +81,7 @@ void arm_stream_decompressor (int32_t command, stream_handle_t instance, stream_
                 memory pointers are in the same order as described in the SWC manifest
         */
         case STREAM_RESET: 
-        {   stream_al_services *stream_entry = (stream_al_services *)(uint64_t)data;
+        {   stream_al_services *stream_entry = (stream_al_services *)data;
             intPtr_t *memresults = (intPtr_t *)instance;
             uint16_t preset = RD(command, PRESET_CMD);
