+++ stream_nodes/Basic/arm/amplifier/arm_stream_amplifier.c
@@ -87,13 +87,13 @@ void arm_stream_amplifier (int32_t command, stream_handle_t instance, stream_xdm
                 memory pointers are in the same order as described in the SWC manifest
         */
         case STREAM_RESET: 
-        {   stream_al_services *stream_entry = (stream_al_services *)(uint64_t)data;
+        {   stream_al_services *stream_entry = (stream_al_services *)data;
             intPtr_t *memresults = (intPtr_t *)instance;
             uint16_t preset = RD(command, PRESET_CMD);
 
             /* here reset */
             arm_amplitude_instance *pinstance = (arm_amplitude_instance *) *memresults;
-            pinstance->services = (stream_al_services *)(uint64_t)data;
+            pinstance->services = (stream_al_services *)data;
 
             memresults++;   /* memresult points to the */
