+++ stream_nodes/Basic/arm/demodulator/arm_stream_demodulator.c
@@ -90,7 +90,7 @@ void arm_stream_demodulator (int32_t command, stream_handle_t instance, stream_x
                 memory pointers are in the same order as described in the SWC manifest
         */
         case STREAM_RESET: 
-        {   stream_al_services *stream_entry = (stream_al_services *)(uint64_t)data;
+        {   stream_al_services *stream_entry = (stream_al_services *)data;
             intPtr_t *memresults = (intPtr_t *)instance;
             uint16_t preset = RD(command, PRESET_CMD);
             uint16_t tag = RD(command, SWC_TAG_CMD);        /* extra parameters */
