+++ stream_nodes/Basic/arm/qos/arm_stream_qos.c
@@ -138,12 +138,12 @@ void arm_stream_qos (int32_t command, stream_handle_t instance, stream_xdmbuffer
                 memory pointers are in the same order as described in the SWC manifest
         */
         case STREAM_RESET: 
-        {   stream_al_services *stream_entry = (stream_al_services *)(uint64_t)data;
+        {   stream_al_services *stream_entry = (stream_al_services *)data;
             intPtr_t *memresults = (intPtr_t *)instance;
             uint16_t preset = RD(command, PRESET_CMD);
 
             arm_qos_instance *pinstance = (arm_qos_instance *) *memresults;
-            pinstance->services = (stream_al_services *)(uint64_t)data;
+            pinstance->services = (stream_al_services *)data;
 
             memresults++;   /* memresult points to the */
