+++ stream_nodes/Basic/arm/resampler/arm_stream_resampler.c
@@ -95,7 +95,7 @@ void arm_stream_resampler (int32_t command, stream_handle_t instance, stream_xdm
                 memory pointers are in the same order as described in the SWC manifest
         */
         case STREAM_RESET: 
-        {   stream_al_services *stream_entry = (stream_al_services *)(uint64_t)data;
+        {   stream_al_services *stream_entry = (stream_al_services *)data;
             intPtr_t *memresults = (intPtr_t *)instance;
             uint16_t preset = RD(command, PRESET_CMD);
